DB: Make model compatible with Hibernate and our SQLite DB

+++ src/main/java/edu/iipw/pap/db/model/PatternStop.java
@@ -21,7 +21,7 @@ public class PatternStop {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "pattern_stop_id")
-    private long patternStopId;
+    private int patternStopId;
 
     @ManyToOne
     @JoinColumn(name = "stop_id", nullable = false)
